The codex alarm is cleared and ark code modification
Description
   The codex alarm is cleared and ark code modification.
Related issue
   #I51RE6:The codex alarm is cleared and ark code modification.

+++ ecmascript/js_api_tree_map_iterator.cpp
@@ -53,7 +53,7 @@ JSTaggedValue JSAPITreeMapIterator::Next(EcmaRuntimeCallInfo *argv)
     }
 
     // Let index be Map.[[NextIndex]].
-    int index = iter->GetNextIndex();
+    int index = static_cast<int>(iter->GetNextIndex());
     if (index < elements) {
         IterationKind itemKind = IterationKind(iter->GetIterationKind());
